shell: don't remove all other panels on desktop panel redraw
panel_configure used to call configure_static_surface which removes all
other surfaces in the panel layer. This unfortunately includes input
panels, such as the one spawned by the keyboard client. As a result,
the keyboard panel was removed on each redraw of the desktop panel, i.e.
each full minute when the clock in the desktop panel changed.